test/requirements.txt
streamlit>=1.25.0
opencv-python-headless>=4.8.0.74
numpy>=1.23.5
Pillow>=10.0.0
